src/src_learner/requirements.txt
numpy==1.22.1
Pillow==9.0.0
pip==21.2.4
setuptools==58.0.4
jpype1==1.3.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
wheel==0.37.1
flask==2.1.1